% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{activity.raw}
\alias{activity.raw}
\title{Activity data (raw)}
\format{A data frame with 17568 rows and 3 variables:
\describe{
  \item{steps}{ Number of steps taking in a 5-minute interval (missing values are coded as NA)}
  \item{date}{The date on which the measurement was taken in YYYY-MM-DD format}
  \item{interval}{Identifier for the 5-minute interval in which measurement was taken}
}}
\usage{
data(activity.raw)
}
\description{
Steps monitoring device data from 2012 as read from CSV file
}
\keyword{datasets}
