% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcess.R
\name{downloadZip}
\alias{downloadZip}
\title{Downloads a zip file, unpacks it on selected folder}
\usage{
downloadZip(fileUrl, outDir = "./data")
}
\arguments{
\item{fileUrl}{the web link to the file}

\item{outDir}{the folder where the file will be placed}
}
\value{
None
}
\description{
Downloads a zip file, unpacks it on selected folder
}

