% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get.imp.df}
\alias{get.imp.df}
\title{Get a complete data frame including the imputed values}
\usage{
get.imp.df(x, mice.imp, uniVar)
}
\arguments{
\item{x}{data frame}

\item{mice.imp}{is an object of mice imputation}

\item{uniVar}{is the variable to look into}
}
\description{
Get a complete data frame including the imputed values
}
