% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{ggplot_missing}
\alias{ggplot_missing}
\title{A function that plots missingness.
   Works only for dataframes with more than 1 variable}
\usage{
ggplot_missing(x)
}
\arguments{
\item{x}{data.frame}
}
\description{
A function that plots missingness.
   Works only for dataframes with more than 1 variable
}

