% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{info.imp}
\alias{info.imp}
\title{Gets a summary statistics for a data frame}
\usage{
info.imp(x, uniVar, colName, decs = 3)
}
\arguments{
\item{x}{data frame}

\item{uniVar}{variable to analyze}

\item{colName}{name of the column in the report table}

\item{decs}{number of decimals. Default is 3.}
}
\value{
data frame with one column
}
\description{
It could be before or after imputation
}

