% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intervals.only}
\alias{intervals.only}
\title{Activity data by intervals with date as rownames}
\format{a data frame derivative of activity}
\usage{
data(intervals)
}
\description{
Intervals are the columns, steps are values. The date is rownames.
Data frame of 61x288. The date as rownames.
}
\keyword{datasets}
