% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intervals}
\alias{intervals}
\title{Activity data by intervals}
\format{An object of class \code{cast_df} (inherits from \code{data.frame}) with 61 rows and 289 columns.}
\usage{
data(intervals)
}
\description{
Intervals are the columns, dates (rows), steps as values.
}
\keyword{datasets}

