% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{nas.plot.frequency}
\alias{nas.plot.frequency}
\title{Plot NAs variable vs frequency}
\usage{
nas.plot.frequency(df, na.var, step = 30, ylim = 0)
}
\arguments{
\item{df}{dataframe}

\item{na.var}{variable with NA}

\item{step}{increment in the breaks}

\item{ylim}{limit y-axis}
}
\description{
Plot NAs variable vs frequency
}
