% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{random.imp}
\alias{random.imp}
\title{Assigns a random value to a vector}
\usage{
random.imp(a)
}
\arguments{
\item{a}{is a vector of completed and NA observations}
}
\value{
a completed vector with random values
}
\description{
The random values are sampled from the completed cases.
}

