% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intervals}
\alias{intervals}
\title{Activity data by intervals}
\format{a data frame derivative of activity}
\usage{
data(intervals)
}
\description{
Intervals are the columns, dates (rows), steps as values.
}
\keyword{datasets}

